Migrate path parameter query param semantic validator to 2And3
