+++ src/plugins/validate-semantic/validators/2and3/paths.js
@@ -1,0 +1,16 @@
+export const validate2And3PathParameterKeysDontContainQuestionMarks = () => system => {
+  return system.validateSelectors
+    .allPathItems()
+    .then(nodes => {
+      return nodes.reduce((acc, node) => {
+        if(node.key.indexOf("?") > -1) {
+          acc.push({
+            message: `Query strings in paths are not allowed.`,
+            path: [...node.path],
+            level: "error",
+          })
+        }
+        return acc
+      }, [])
+    })
+}
