+++ src/plugins/validate-semantic/validators/paths.js
@@ -82,24 +82,6 @@ export const validatePathParameterKeysAreDifferent = () => system => {
     })
 }
 
-export const validatePathParameterKeysDontContainQuestionMarks = () => system => {
-  return system.validateSelectors
-    .allPathItems()
-    .then(nodes => {
-      return nodes.reduce((acc, node) => {
-        if(node.key.indexOf("?") > -1) {
-          acc.push({
-            message: `Query strings in paths are not allowed.`,
-            path: [...node.path],
-            level: "error",
-          })
-        }
-        return acc
-      }, [])
-    })
-}
-
-
 /// Helpers
 
 function checkForDefinition(paramName, pathItem) {
